utils.py
import yaml

def read_config(config_filename):

    with open(config_filename, 'r') as stream:
        try:
            config = yaml.load(stream, Loader=yaml.FullLoader)
        except yaml.YAMLError as exc:
            print(exc)

    return config

def save_config(data, config_filename):
    with open(config_filename, 'w') as stream:
        try:
            yaml.dump(data, stream)
        except yaml.YAMLError as exc:
            print(exc)
