requirements.txt
argparse==1.4.0
biornn==1.0
celluloid==0.2.0
dm-sonnet==2.0.0
pip-chill==1.0.1
pyyaml==5.3.1
tensorflow==2.3.0
